feat: Show total sales of Event (#207)

+++ app/src/main/java/org/fossasia/openevent/app/data/models/Order.java
@@ -13,7 +13,7 @@ public class Order extends BaseModel {
     @PrimaryKey
     private long id;
 
-    private long amount;
+    private float amount;
     @SerializedName("completed_at")
     private String completedAt;
     private String identifier;
@@ -31,11 +31,11 @@ public class Order extends BaseModel {
         setIdentifier(identifier);
     }
 
-    public long getAmount() {
+    public float getAmount() {
         return amount;
     }
 
-    public void setAmount(long amount) {
+    public void setAmount(float amount) {
         this.amount = amount;
     }
